navbar fix
The navbar in the admin app should now use admin/base/nav-bar.html for
the top navbar. This way when in the admin section we can see the
links and stuff to deal with stores, charts, users, etc etc

The problem was you couldnt see nav items on mobile before. We think
we have fixed that.